Refactor utilities to core/utils.js and enable ES modules
Moved shared utility functions (hashId, inferFromCategory, fmtDate, normalizeStr, jaccard, tokenSet) to a new core/utils.js module. Updated background.js, dashboard.js, and manifest.json to use ES module imports. Changed script tags in dashboard.html and options.html to use type="module". Improved WebSocket reconnection logic with exponential backoff. Enhanced server status display in options.js with status icons.
