+++ core/utils.js
@@ -1,0 +1,58 @@
+// Shared utility functions for Marcalink Snowballing
+
+export function fmtDate(iso) {
+    if (!iso) return "";
+    try {
+        const d = new Date(iso);
+        return d.toLocaleDateString("pt-BR");
+    } catch {
+        return iso;
+    }
+}
+
+export function normalizeStr(s) {
+    return (s || "").toString().toLowerCase();
+};
+
+export function tokenSet(title) {
+    return new Set(utils.normalizeStr(title)
+        .replace(/[^a-z0-9\s]/g, " ")
+        .split(/\s+/)
+        .filter(w => w && w.length >= 3));
+};
+
+export function jaccard(a, b) {
+    const A = utils.tokenSet(a);
+    const B = utils.tokenSet(b);
+    if (!A.size || !B.size) return 0;
+    let inter = 0;
+    for (const x of A) if (B.has(x)) inter++;
+    const uni = A.size + B.size - inter;
+    return uni ? inter / uni : 0;
+};
+
+// FNV-1a 32-bit hash, prefixed with p_
+export function hashId(input) {
+    input = (input || "").toString();
+    let h = 0x811c9dc5;
+    for (let i = 0; i < input.length; i++) {
+        h ^= input.charCodeAt(i);
+        h = (h * 0x01000193) >>> 0;
+    }
+    return "p_" + h.toString(16).padStart(8, "0");
+};
+
+export function inferFromCategory(category) {
+    const c = (category || "").toLowerCase();
+    const origin = c.includes("seed") || c.includes("semente") ? "seed"
+        : c.includes("back") || c.includes("refer") ? "backward"
+        : c.includes("forw") || c.includes("cita") ? "forward"
+        : "unknown";
+
+    const status = c.includes("incl") ? "included"
+        : c.includes("excl") ? "excluded"
+        : c.includes("duplic") ? "duplicate"
+        : "pending";
+
+    return { origin, status };
+};
