+++ dashboard.html
@@ -302,6 +302,6 @@
     </div>
 
     <script src="svat_storage.js"></script>
-    <script src="dashboard.js"></script>
+    <script type="module" src="dashboard.js"></script>
   </body>
 </html>
