+++ options.html
@@ -5,7 +5,7 @@
     <meta name="viewport" content="width=device-width, initial-scale=1" />
     <title>Configurações • Marcalink Snowballing</title>
     <link rel="stylesheet" href="options.css" />
-    <script defer src="options.js"></script>
+    <script defer type="module" src="options.js"></script>
   </head>
   <body>
     <div class="layout">
